backend/requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.18
django-colorfield==0.8.0
django-cors-headers==3.14.0
django-extra-fields==3.0.2
django-import-export==3.1.0
django-templated-mail==1.1.1
django-weasyprint==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
# fonttools==4.39.3
html5lib==1.1
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.4.0
pipreqs==0.4.11
# psycopg2-binary
pycparser==2.21
pydyf==0.6.0
PyJWT==2.6.0
pyphen==0.14.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
reportlab==3.6.13
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
tablib==3.4.0
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
weasyprint==58.1
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
zipp==3.15.0
zopfli==0.2.2
